Rename GoogleMaps scripts to D- instead of E-
