+++ bus_service_increase/D2_setup_gmaps.py
@@ -13,7 +13,7 @@ from datetime import timedelta
 from calitp.storage import get_fs
 
 from bus_service_utils import utils
-from E1_setup_parallel_trips_with_stops import ANALYSIS_DATE, COMPILED_CACHED
+from D1_setup_parallel_trips_with_stops import ANALYSIS_DATE, COMPILED_CACHED
 
 fs = get_fs()
 catalog = intake.open_catalog("./*.yml")
